+++ server/testingHelpers.js
@@ -102,6 +102,7 @@ function sessionFactory(overrides = {}) {
     overrides.respondedAt !== undefined ? overrides.respondedAt : new Date('2021-10-05T20:20:33.000Z'),
     overrides.respondedByPhoneNumber !== undefined ? overrides.respondedByPhoneNumber : null,
     overrides.location !== undefined ? overrides.location : locationFactory(),
+    overrides.numberOfAlerts !== undefined ? overrides.numberOfAlerts : 1,
   )
 }