+++ CHANGELOG.md
@@ -11,6 +11,12 @@ the code was deployed.
 
 ## [Unreleased]
 
+### Added
+
+- `number_of_alerts` integer column to the sessions table (CU-860r8k57h).
+- Chatbot flow to reset a sensor if the `number_of_alerts` in a session exceeds the threshold `SESSION_NUMBER_OF_ALERTS_TO_ACCEPT_RESET_REQUEST` (CU-860r8k57h).
+- English and (temporary) Spanish translations of `alertAcceptResetRequest`, `clientMessageForRequestToReset`, `resetNoticeToRequester`, `resetNoticeToOtherResponders`, `resetRequestRejected` (CU-860r8k57h).
+
 ## [10.5.0] - 2024-01-08
 
 ### Removed